implemented peer check functionality
